feat: Add RecommendationStats model and update recycling-location model
This commit adds a new model called RecommendationStats, which represents the statistics and recommendations for a user. It also updates the recycling-location model to include a new field called recyclingCenterDistances, which stores the distances between the recycling location and other recycling centers. These changes enhance the functionality of the application by providing recommendations and improving the representation of recycling locations.